requirements.txt
pymongo[srv]
python-dotenv
requests
urllib3==1.26.16
jinja2
certifi
